refactor: move app.py into package root
