--- src/harlequin/tui/__init__.py
@@ -1,1 +1,0 @@
-from harlequin.tui.app import Harlequin as Harlequin
