+++ tests/functional_tests/conftest.py
@@ -2,7 +2,7 @@ import shutil
 from pathlib import Path
 
 import pytest
-from harlequin.tui import Harlequin
+from harlequin import Harlequin
 
 
 @pytest.fixture
